Server Application v1.0
UI and code is fixed for the first pre-build.

Files have been reorganised and code paths updated to work for the deployment. This means that there are duplicated files.

There is the main directory for changes to be made, these are copied into the Qt Files/deploy for testing.

Hidden folder in each for deploy/building the code.,